pain of serializable recipes+fftontmf..

+++ src/main/java/com/hbm/entity/projectile/EntityArtilleryShell.java
@@ -261,8 +261,8 @@ public class EntityArtilleryShell extends EntityThrowableNT implements IChunkLoa
     }
 
     @Override
-    public float getGravityVelocity() {
-        return 9.81F * 0.05F;
+    public double getGravityVelocity() {
+        return 9.81 * 0.05;
     }
 
     @Override
